coalesce all utils in wsi/utils.py; less exposed methods; more docstrings
