--- wsi/wsi_dataset.py
@@ -1,8 +1,0 @@
-from torchvision import transforms
-
-
-def default_transforms(mean=(0.485, 0.456, 0.406), std=(0.229, 0.224, 0.225)):
-    t = transforms.Compose(
-        [transforms.ToTensor(), transforms.Normalize(mean=mean, std=std)]
-    )
-    return t
