--- wsi/wsi_utils.py
@@ -1,72 +1,0 @@
-import h5py
-import numpy as np
-
-
-def save_hdf5(output_path, asset_dict, attr_dict=None, mode="a"):
-    file = h5py.File(output_path, mode)
-    for key, val in asset_dict.items():
-        data_shape = val.shape
-        if key not in file:
-            data_type = val.dtype
-            chunk_shape = (1,) + data_shape[1:]
-            maxshape = (None,) + data_shape[1:]
-            dset = file.create_dataset(
-                key,
-                shape=data_shape,
-                maxshape=maxshape,
-                chunks=chunk_shape,
-                dtype=data_type,
-            )
-            dset[:] = val
-            if attr_dict is not None:
-                if key in attr_dict.keys():
-                    for attr_key, attr_val in attr_dict[key].items():
-                        dset.attrs[attr_key] = attr_val
-        else:
-            dset = file[key]
-            dset.resize(len(dset) + data_shape[0], axis=0)
-            dset[-data_shape[0] :] = val
-    file.close()
-    return output_path
-
-
-def sample_indices(scores, k, start=0.48, end=0.52, convert_to_percentile=False, seed=1):
-    np.random.seed(seed)
-    if convert_to_percentile:
-        end_value = np.quantile(scores, end)
-        start_value = np.quantile(scores, start)
-    else:
-        end_value = end
-        start_value = start
-    score_window = np.logical_and(scores >= start_value, scores <= end_value)
-    indices = np.where(score_window)[0]
-    if len(indices) < 1:
-        return -1
-    else:
-        return np.random.choice(indices, min(k, len(indices)), replace=False)
-
-
-def top_k(scores, k, invert=False):
-    if invert:
-        top_k_ids = scores.argsort()[:k]
-    else:
-        top_k_ids = scores.argsort()[::-1][:k]
-    return top_k_ids
-
-
-def to_percentiles(scores):
-    from scipy.stats import rankdata
-
-    scores = rankdata(scores, "average") / len(scores) * 100
-    return scores
-
-
-def screen_coords(scores, coords, top_left, bot_right):
-    bot_right = np.array(bot_right)
-    top_left = np.array(top_left)
-    mask = np.logical_and(
-        np.all(coords >= top_left, axis=1), np.all(coords <= bot_right, axis=1)
-    )
-    scores = scores[mask]
-    coords = coords[mask]
-    return scores, coords
